+++ lib/lightning/usage_tracking/report.ex
@@ -5,6 +5,8 @@ defmodule Lightning.UsageTracking.Report do
   """
   use Ecto.Schema
 
+  import Ecto.Changeset
+
   @primary_key {:id, :binary_id, autogenerate: true}
   schema "usage_tracking_reports" do
     field :data, :map
@@ -13,5 +15,12 @@ defmodule Lightning.UsageTracking.Report do
     field :report_date, :date
 
     timestamps()
+  end
+
+  def changeset(report, params) do
+    report
+    |> cast(params, [:data, :report_date, :submitted, :submitted_at])
+    |> validate_required([:data, :report_date, :submitted])
+    |> unique_constraint(:report_date)
   end
 end